requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.35.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
firebase-admin==6.5.0
python-dateutil==2.8.2
